# docs/api-openapi.yaml
openapi: 3.0.3
info:
  title: QuicK ReeLS API
  version: 0.1.0
  description: >
    Business API used by the TikTok Mini H5 and the separate operations console.
    Provider credentials and raw media URLs never leave the server.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production placeholder
...
paths:
...
  /episodes/{episodeId}/play:
    get:
      tags: [Playback]
      summary: Get the minimum official-player playback DTO
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/EpisodeId'
      responses:
        '200':
          description: Platform playback identifiers and resume position
        '403':
          description: Episode is locked
        '409':
          description: Platform playback mapping is not ready
  /episodes/{episodeId}/reward-unlock:
    post:
      tags: [Playback]
      summary: Idempotently unlock an episode after rewarded ad completion
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/EpisodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [placementId, clientEventId, isEnded]
              properties:
                placementId: { type: string }
                clientEventId: { type: string, maxLength: 128 }
                isEnded: { type: boolean, enum: [true] }
                sessionId: { type: string }
      responses:
        '200':
          description: Episode is playable
...
  /me/history:
    get:
      tags: [User session]
      summary: Get the current user's viewing history
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: History records
  /me/favorites:
    get:
      tags: [User session]
      summary: Get the current user's favorites
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Favorite albums
  /me/preferences:
    get:
      tags: [User session]
      summary: Get preferences
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Locale and playback preferences
    put:
      tags: [User session]
      summary: Update preferences
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Updated preferences
...
components:
  parameters:
...
    EpisodeId:
      in: path
      name: episodeId
      required: true
      schema: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminBearerAuth:
      type: http
      scheme: bearer